Added Configuration Logic
This logic does not currently send a valid json request to the Slang
frontend-- but-- this is a framework where things are mostly working.
All settings are tracked; when project settings are changed all open project LSP servers are
updated accordingly.

The remaining issue is that Slang does not
respect the JSON query sent to the LSP server.